@@ -5,7 +5,7 @@
     <SectionPanel
       name="Colour Palette"
       :open="openPanels.includes('color')"
-      @update:open="updatePanelState({ panel: 'color', open: $event})"
+      @update:open="updatePanelState({ panel: 'color', open: $event })"
     >
       <ColorPalette />
     </SectionPanel>
@@ -13,7 +13,7 @@
     <SectionPanel
       name="Dimensions"
       :open="openPanels.includes('dimensions')"
-      @update:open="updatePanelState({ panel: 'dimensions', open: $event})"
+      @update:open="updatePanelState({ panel: 'dimensions', open: $event })"
     >
       <div class="row">
         <DimensionAdjuster />
@@ -24,7 +24,7 @@
     <SectionPanel
       name="General Settings"
       :open="openPanels.includes('general')"
-      @update:open="updatePanelState({ panel: 'general', open: $event})"
+      @update:open="updatePanelState({ panel: 'general', open: $event })"
     >
       <GeneralSettings />
     </SectionPanel>
@@ -32,7 +32,7 @@
     <SectionPanel
       name="Knitting"
       :open="openPanels.includes('knitting')"
-      @update:open="updatePanelState({ panel: 'knitting', open: $event})"
+      @update:open="updatePanelState({ panel: 'knitting', open: $event })"
     >
       <div class="row">
         <KnitSession />
@@ -43,7 +43,7 @@
     <SectionPanel
       name="Load &amp; Import"
       :open="openPanels.includes('loadImport')"
-      @update:open="updatePanelState({ panel: 'loadImport', open: $event})"
+      @update:open="updatePanelState({ panel: 'loadImport', open: $event })"
     >
       <LoadImportPanel />
     </SectionPanel>
@@ -51,7 +51,9 @@
 </template>
 
 <script>
-import { mapActions, mapGetters } from 'vuex';
+import { storeToRefs } from 'pinia';
+import { useSettingsStore } from '@/store/settings/state';
+import { usePersistanceStore } from '@/store/persistance/state';
 import PrimaryInfo from './controls/PrimaryInfo.vue';
 import ColorPalette from './controls/ColorPalette.vue';
 import GeneralSettings from './controls/GeneralSettings.vue';
@@ -74,17 +76,22 @@ export default {
     LoadImportPanel,
     SectionPanel,
   },
-  computed: {
-    ...mapGetters(['openPanels']),
+  setup() {
+    const settingsStore = useSettingsStore();
+    const { openPanels } = storeToRefs(settingsStore);
+    const { updatePanelState } = settingsStore;
+
+    const persistanceStore = usePersistanceStore();
+    const { loadSession } = persistanceStore;
+
+    return {
+      openPanels,
+      loadSession,
+      updatePanelState,
+    };
   },
   created() {
     this.loadSession();
-  },
-  methods: {
-    ...mapActions([
-      'loadSession',
-      'updatePanelState',
-    ]),
   },
 };
 </script>
@@ -95,13 +102,16 @@ export default {
   flex-direction: row;
   align-items: stretch;
 }
-.row > * {
+
+.row>* {
   flex: 1;
 }
+
 .grid-controls {
   overflow-y: auto;
   flex: 1;
 }
+
 section:last-child {
   border-bottom: 2px solid grey;
 }
